requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Translation helpers
django-modeltranslation>=0.19  # Translate model fields
django-parler>=2.3             # Alternative for translatable models
djangorestframework-simplejwt>=5.3  # (optional) if you want auth while testing APIs
django-rosetta>=0.9.0  # For managing translations via UI
